+++ docs/changelog.md
@@ -1,11 +1,15 @@
 # Changelog
 
-## [v0.2.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.3) - Upcoming Release
+## [v0.2.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.3) - Mar 14 2024
 
  - Updated config management
  - Improved config file validation
  - Fixed issue where HTTP requests would fail if only IP & port were specified in config
  - Fixed crash if Frigate has no event info & app couldn't pull snapshots
+ - Added ability to filter notifications based on [zones](https://frigate-notify.0x2142.com/config/#zones)
+     - New config option to only allow notifications based on zone
+     - New config option to deny notifications based on zone
+     - New config option to allow or drop events outside of a designated zone
 
 ## [v0.2.2](https://github.com/0x2142/frigate-notify/releases/tag/v0.2.2) - Nov 29 2023